+++ src/app/service/map-id.service.ts
@@ -1,0 +1,14 @@
+import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+
+@Injectable({
+	providedIn: 'root'
+})
+export class MapIdService {
+
+	constructor(private http: HttpClient) { }
+
+	getMapId() {
+		return this.http.get("mapid.info");
+	}
+}