+++ src/environments/environment.public.ts
@@ -1,0 +1,5 @@
+export const environment = {
+	production: true,
+	mapTileUrl: "https://{s}.omega-zirkel.de/{mapId}/{z}/{x}/{y}.png",
+	mapTileSubdomains: ['xtiles01']
+};